versionName "1.2.3"  尝试修复报错
 error: non-constant-expression cannot be narrowed from type 'int' to 'SLuint32'

+++ qiaopcplayer/src/main/cpp/MyAudio.cpp
@@ -377,7 +377,7 @@ int MyAudio::getCurrentSampleRateForOpensles(int sample_rate) {
         default:
             rate =  SL_SAMPLINGRATE_44_1;
     }
-    return rate;
+    return (uint)rate;
 }
 
 void MyAudio::pause() {
